requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
awscrt==0.14.0
backoff==2.1.2
boto3[crt]==1.24.65
botocore==1.27.65
certifi==2022.6.15
charset-normalizer==2.1.1
coloredlogs==15.0.1
docopt==0.6.2
-e .
frozenlist==1.3.1
gql[all]==3.4.0
graphql-core==3.2.1
humanfriendly==10.0
idna==3.3
jmespath==1.0.1
lizard==1.17.10
multidict==6.0.2
py-bip39-bindings==0.1.9
python-dateutil==2.8.2
requests==2.28.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
setuptools==65.3.0
six==1.16.0
urllib3==1.26.12
websockets==10.3
xmltodict==0.13.0
yarl==1.8.1
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
glob2==0.7
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
jinja2==3.1.2
json5==0.9.10
jsonschema==4.15.0
jupyter-client==7.3.5
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.1
lxml==4.9.1
mako==1.2.2
markupsafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
packaging==21.3
pandocfilters==1.5.0
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-bdd==6.0.1
pytest-dependency==0.5.1
pytz==2022.2.1
pyzmq==23.2.1
send2trash==1.8.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
zipp==3.8.1
